Refactor admin.js and make db and email helpers files
