+++ server/helpers.js
@@ -1,3 +1,5 @@
+const crypto = require('crypto');
+
 const getToday = () => {
   let today = new Date();
   let dd = today.getDate();
@@ -23,8 +25,28 @@ const trace = label => x => {
   return x;
 };
 
+const generate_random_token = () => {
+  return new Promise((resolve, reject) => {
+    crypto.randomBytes(20, (err, buffer) => {
+      if (err) {
+        console.log('Error generating random token: ', err);
+        return reject(err);
+      }
+      const token = buffer.toString('hex');
+      return resolve(token);
+    });
+  });
+};
+
+const parse_date = date => {
+  if (!date) return 0;
+  return Date.parse(date);
+};
+
 module.exports = {
   getToday,
   randomIntegerInRange,
-  trace
+  trace,
+  generate_random_token,
+  parse_date
 };
